docs/istio/File_upload_gcp_and_reservation_service/requirements.txt
djangorestframework==3.11.0
Django==3.0.4
pymongo==3.10.1
requests==2.23.0
mongoengine==0.19.1
prometheus_client==0.7.1
pytest
coverage
django-allauth
google-cloud
pandas
google-cloud-storage